Fix code review issues and break circular import cycles
- Fix CLAUDE.md class names (ScoutGraphContext → FindConsiderationsGraphContext,
  MultiRoundScoutLoop → MultiRoundLoop)
- Make CallRunner inherit from ABC so @abstractmethod is enforced
- Remove unnecessary f-string prefixes on lines without placeholders
- Propagate Sequence through type hints instead of using list where
  mutation isn't needed (parameters in llm.py, common.py, context.py,
  stages.py, closing_reviewers.py, context_builders.py)
- Extract assess_concept_types.py to break closing_reviewers ↔
  assess_concept circular import
- Move link_new_pages() to context_builders.py to break
  context_builders ↔ find_considerations circular import
- Promote all deferred imports in calls/ and orchestrator.py to
  top-level now that cycles are resolved
- Move FruitCheck model to module level in page_creators.py
